extracting controller logic
